rename: calendarModal, calendar 컴포넌트 각각 calendar, month로 이름 변경
